Adds Junit test for AddDeadlineCommand
