+++ src/test/java/ip/commands/AddDeadlineCommandTest.java
@@ -1,0 +1,43 @@
+package ip.commands;
+
+import ip.exceptions.FileCorruptedException;
+import ip.exceptions.UnknownInputException;
+import ip.storage.Storage;
+import ip.storage.StorageStub;
+import ip.tasks.Deadline;
+import ip.tasks.Task;
+import ip.tasks.TaskList;
+import ip.ui.Ui;
+import org.junit.jupiter.api.Test;
+
+import java.time.LocalDate;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class AddDeadlineCommandTest {
+    @Test
+    public void execute_success() throws UnknownInputException, FileCorruptedException {
+        AddDeadlineCommand testCommand = new AddDeadlineCommand();
+        Ui ui = new Ui();
+        Storage storageStub = new StorageStub();
+
+        TaskList testListOne = new TaskList();
+        testCommand.execute("deadline homework /by 2025-01-01", ui, storageStub, testListOne);
+        Task testOne = testListOne.get(0);
+        assertEquals(new Deadline("homework", LocalDate.parse("2025-01-01")).toString(),
+                testOne.toString());
+
+        TaskList testListTwo = new TaskList();
+        testCommand.execute("Deadline wash the clothes /by 2022-10-15", ui, storageStub, testListTwo);
+        Task testTwo = testListTwo.get(0);
+        assertEquals(new Deadline("wash the clothes", LocalDate.parse("2022-10-15")).toString(),
+                testTwo.toString());
+
+        TaskList testListThree = new TaskList();
+        testCommand.execute("deadline Finish 2030 Ip /by 2026-08-31", ui, storageStub, testListThree);
+        Task testThree = testListThree.get(0);
+        assertEquals(new Deadline("Finish 2030 Ip", LocalDate.parse("2026-08-31")).toString(),
+                testThree.toString());
+
+    }
+}
