+++ src/test/java/ip/tasks/TodoTest.java
@@ -1,0 +1,17 @@
+package ip.tasks;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class TodoTest {
+    @Test
+    public void dummyTest() {
+        assertEquals(20, 20);
+    }
+
+    @Test
+    public void anotherDummyTest() {
+        assertEquals(40, 40);
+    }
+}
